Session 2/LinkedIn/requirements.txt
# Required Python packages for LinkedIn Job Tracker

# Google Sheets API
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0

# Web scraping and parsing
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3

# Email
secure-smtplib==0.1.1

# Date and time handling
python-dateutil==2.8.2
pytz==2023.3

# Scheduling
schedule==1.2.0

# Data handling
pandas==2.1.4
openpyxl==3.1.2
